FN-Scrapers-postgres_insert/scrapers/requirements.txt
-e git+git@github.com:FiscalNote/FN-DataAccess-Service@2.4.8#egg=fn_dataaccess_client&subdirectory=client_python
-e git+git@github.com:FiscalNote/FN-DocumentService@1.5.0#egg=fn_document_service&subdirectory=thrift_gen_python
-e git+git@github.com:FiscalNote/FN-Logging-Python.git@2.1.3#egg=fn_logging
-e git+git@github.com:FiscalNote/FiscalNote-Rabbit-Python.git@1.3.4#egg=fn_rabbit
-e git+git@github.com:FiscalNote/FiscalNote-RateLimiter.git@3.1.0#egg=fn_ratelimiter_client&subdirectory=client
-e git+git@github.com:FiscalNote/FiscalNote-RateLimiter.git@3.1.0#egg=fn_ratelimiter_common&subdirectory=common
-e git+git@github.com:FiscalNote/FN-ScraperUtils.git@0.8.29#egg=fn_scraperutils
-e git+git@github.com:FiscalNote/FN-Service-Python@0.8.5#egg=fn_service
-e git+git@github.com:abhiatoda/python-jsonschema-objects.git@0.3.1#egg=python_jsonschema_objects
-e git+ssh://git@github.com/FiscalNote/FN-Mapper.git@v1.0#egg=fn_mapper

anyconfig==0.9.1
Babel==2.5.0
dateparser==0.6.0
inflection==0.2.0
pycountry==17.5.14
simplejson==3.11.1
suds-jurko==0.6
w3lib==1.18.0
xmltodict==0.11.0
Markdown==2.4
regex==2017.7.28
tzlocal==1.4
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.34

PyVirtualDisplay==0.2.1
EasyProcess==0.2.3
selenium==3.6.0

arrow==0.10.0
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
boto==2.48.0
certifi==2017.4.17
cffi==1.10.0
constantly==15.1.0
cryptography==1.9
enum34==1.1.6
functools32==3.2.3.post2
future==0.16.0
futures==3.1.1
httplib2==0.10.3
idna==2.5
incremental==17.5.0
injector==0.12.1
ipaddress==1.0.18
jsonschema==2.6.0
lxml==4.1.1
ndg-httpsclient==0.4.2
oauth2==1.9.0.post1
pika==0.10.0
psycopg2==2.6.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
pylru==1.0.9
pyOpenSSL==17.0.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.14.2
rfc3987==1.3.7
service-identity==17.0.0
six==1.10.0
SQLAlchemy==1.0.10
strict-rfc3339==0.7
subprocess32==3.2.7
thrift==0.10.0
tinys3==0.1.12
hyperlink==17.3.1
Twisted==17.9.0
typing==3.6.1
zope.interface==4.4.1
argcomplete==1.8.2
cssselect==1.0.1
parsel==1.2.0
Unidecode==0.4.20
alembic==0.9.2
Mako==1.0.6
MarkupSafe==1.0
python-editor==1.0.3
croniter==0.3.17
ansicolors==1.1.8
xlrd==1.0.0

numpy==1.13.1
pandas==0.20.3
scikit-learn==0.18.2
scipy==0.19.1

html5lib==0.999999999
webencodings==0.5.1
pysqlite==2.8.3
pyzipcode==1.0
mechanize==0.2.5

# ks_bills
ezodf==0.3.2

# common/files
LEPL==5.1.3
rfc6266==0.0.4

# RFC 3986 complaint url utility
uritools==2.2.0

# common/json_consumer
jsonpath_ng==1.4.3
decorator==4.3.0
ply==3.11
parsedatetime==2.4

# CT EventScraper
csv23==0.1.4
